Fix filter values not returning to defaults on image load
When the image is first loaded, no filters are applied, thus their initial value must be zero. The default value when not set on the HTML element was 50, so those were adjusted as well.

I saw that you labeled the canvas brightness filter for exposure. To get around the mismatches without modifying your code, I added a data attribute where you can specify the filter name.